+++ releasenotes.md
@@ -3,6 +3,7 @@
 ## 0.5.0 Change List
 
 ### Changes / improvements
+- '$$select' builtin for vector masked select.
 - Subtype matching in type switches.
 - Added parentof typeid property.
 - Slice assignment is expanded.